✨ スライドデータ管理用サーバーサイドコードの作成
