+++ server/api/slides.ts
@@ -1,28 +1,5 @@
-import { SlideInfo } from "~/model/slide";
-
-import { createClient } from "microcms-js-sdk";
-
-const config = useRuntimeConfig();
-
-const client = createClient({
-  serviceDomain: config.microCMSServiceDomain,
-  apiKey: config.microCMSApiKey,
-});
-
-const data = [] as SlideInfo[];
-
-client
-  .getList<SlideInfo>({
-    endpoint: "slide",
-    queries: { limit: 100 },
-  })
-  .then((res) => {
-    console.log(res);
-    res.contents.forEach((slide) => {
-      data.push(slide);
-    });
-  });
+import { getSlides } from "~/server/slides";
 
 export default defineEventHandler((event) => {
-  return data;
+  return getSlides();
 });
